backend/requirements.txt
langchain
pinecone-client
openai
python-dotenv